autografs_requirements.txt
AuToGraFS
ase
pyyaml
